Put C API behind feature gate
