+++ rust/template/differential_datalog/int.rs
@@ -12,9 +12,11 @@ use ordered_float::OrderedFloat;
 use serde::de::Error;
 use serde::de::*;
 use serde::ser::*;
+#[cfg(feature = "c_api")]
 use std::ffi::CStr;
 use std::fmt;
 use std::ops::*;
+#[cfg(feature = "c_api")]
 use std::os::raw::c_char;
 use std::str::FromStr;
 
@@ -157,22 +159,26 @@ impl Int {
 }
 
 #[no_mangle]
+#[cfg(feature = "c_api")]
 pub extern "C" fn int_from_i64(v: i64) -> *mut Int {
     Box::into_raw(Box::new(Int::from_i64(v)))
 }
 
 #[no_mangle]
+#[cfg(feature = "c_api")]
 pub extern "C" fn int_from_u64(v: u64) -> *mut Int {
     Box::into_raw(Box::new(Int::from_u64(v)))
 }
 
 #[no_mangle]
+#[cfg(feature = "c_api")]
 pub unsafe extern "C" fn int_from_str(s: *const c_char, radix: u32) -> *mut Int {
     let c_str = CStr::from_ptr(s);
     Box::into_raw(Box::new(Int::parse_bytes(c_str.to_bytes(), radix)))
 }
 
 #[no_mangle]
+#[cfg(feature = "c_api")]
 pub unsafe extern "C" fn int_free(x: *mut Int) {
     if x.is_null() {
         return;
