+++ rust/template/src/ovsdb_api.rs
@@ -1,5 +1,5 @@
 //! OVSDB JSON interface to RunningProgram
-#![cfg(feature = "ovsdb")]
+#![cfg(all(feature = "ovsdb", feature = "c_api"))]
 
 use std::collections::BTreeMap;
 use std::convert::TryFrom;
